+++ sys/net/ipfw/ip_fw2.c
@@ -2946,7 +2946,8 @@ ipfw_flush(int kill_default)
 			("%u static rules remains\n", static_count));
 		KASSERT(static_ioc_len == IOC_RULESIZE(ctx->ipfw_default_rule),
 			("%u bytes of static rules remains, should be %lu\n",
-			 static_ioc_len, IOC_RULESIZE(ctx->ipfw_default_rule)));
+			 static_ioc_len,
+			 (u_long)IOC_RULESIZE(ctx->ipfw_default_rule)));
 	}
 
 	/* Flush is done */
